Move image synthesis to sampling/images.py; drop mode-specific temperature
- Relocate build_synthetic_image_uris/parse_image_size out of the CLI module
  into tokenomics/sampling/images.py — synthetic content generation belongs
  with the text samplers, not in benchmark orchestration. Exported from the
  sampling package.
- Temperature no longer varies by mode (was 0 for image runs, 0.7 otherwise):
  temperature doesn't affect a speed benchmark, so the dual default was
  needless complexity. Single default 0.7; pass --temperature 0 / --ignore-eos
  for reproducible generation. (Image output still defaults to 32 tokens.)
